+++ .gitignore
@@ -1,3 +1,6 @@
+
+
+*openai-keys.env 
 **/*.ipynb_checkpoints/
 
 .DS_Store
