+++ local_server.bat
@@ -2,7 +2,7 @@
 :: Licensed under the MIT License.
 
 @echo off
-set FLASK_APP=app.py
+set FLASK_APP=py-src/data_formulator/app.py
 set FLASK_RUN_PORT=5000
 set FLASK_RUN_HOST=0.0.0.0
 flask run
