+++ py-src/data_formulator/__init__.py
@@ -1,0 +1,5 @@
+from .app import run_app
+
+__all__ = [
+    "run_app",
+]
