+++ py-src/data_formulator/__main__.py
@@ -1,0 +1,4 @@
+from .app import run_app
+
+if __name__ == "__main__":
+    run_app()
