+++ py-src/data_formulator/agents/agent_code_explanation.py
@@ -2,7 +2,7 @@
 # Licensed under the MIT License.
 
 import pandas as pd
-from agents.agent_utils import generate_data_summary, extract_code_from_gpt_response
+from data_formulator.agents.agent_utils import generate_data_summary, extract_code_from_gpt_response
 
 import logging
