+++ py-src/data_formulator/agents/agent_data_rec.py
@@ -3,11 +3,13 @@
 
 import json
 
-from agents.agent_utils import extract_json_objects, generate_data_summary, extract_code_from_gpt_response
-import py_sandbox
+from data_formulator.agents.agent_utils import extract_json_objects, generate_data_summary, extract_code_from_gpt_response
+from data_formulator.agents.agent_data_transform_v2 import completion_response_wrapper
+
+import data_formulator.py_sandbox as py_sandbox
+
 import traceback
 
-from agents.agent_data_transform_v2 import completion_response_wrapper
 
 import logging
