--- server/agents/__init__.py
@@ -1,22 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# Licensed under the MIT License.
-
-from .agent_concept_derive import ConceptDeriveAgent
-from .agent_py_concept_derive import PyConceptDeriveAgent
-from .agent_data_transformation import DataTransformationAgent
-from .agent_data_transform_v2 import DataTransformationAgentV2
-from .agent_data_load import DataLoadAgent
-from .agent_sort_data import SortDataAgent
-from .agent_data_clean import DataCleanAgent
-from agents.agent_data_rec import DataRecAgent
-
-__all__ = [
-    "ConceptDeriveAgent",
-    "PyConceptDeriveAgent",
-    "DataTransformationAgent",
-    "DataTransformationAgentV2",
-    "DataRecAgent",
-    "DataLoadAgent",
-    "SortDataAgent",
-    "DataCleanAgent"
-]
