+++ vite.config.ts
@@ -9,4 +9,7 @@ export default defineConfig({
       '@': path.resolve(__dirname, './src'),
     },
   },
+  build: {
+    outDir: path.join(__dirname, 'py-src', 'data_formulator', "dist"),
+  },
 });
